Get overloaded function IPE while avoiding joins
We stopped passing the overloaded function to the wrapper when we realized we
needed to handle joins, but I don't think we needed to. We now do this again,
but avoid instrumenting join points.

This works well! We also generate uniques for each instrumentation point and
that seems useful so far.

We have reached a point where we need a better way to report the overloaded
calls, as the instrumented cabal-install executable fails while waiting to write
to the file. Database? Eventlog?