Workaround blackholed dicts, better CC names
When the `Cabal-syntax` lib is instrumented by the plugin and then the cabal
hackage-tests parsec test is ran, many `CharParsing ParsecParser` dictionaries
come through as black hole closures, making them invisible to the plugin. We now
force the dict and retry in the black hole case, which appears to fix the issue.

Also do not include call ids in the CC names added by the plugin.